refactor: Attack Chain Rework (#5672)
* Attack Chain Rework

* Laser Eyes Attack Fix

* No Duplicates

* Pipe Turf Check

* Gibtonite Addition

* Roboquest Console Update

* Buckled Null Fix

* Attack Results Update

* Cleaver Update

* Utencil Overlay Fix

+++ code/game/objects/structures/pit.dm
@@ -18,32 +18,58 @@
 /obj/structure/pit/AllowDrop()
     return TRUE
 
-/obj/structure/pit/attackby(obj/item/W, mob/user)
-	if(istype(W,/obj/item/shovel))
-		visible_message("<span class='notice'>\The [user] starts [open ? "filling" : "digging open"] \the [src]</span>")
-		if(do_after(user, 5 SECONDS * W.toolspeed, src, category = DA_CAT_TOOL))
-			visible_message("<span class='notice'>\The [user] [open ? "fills" : "digs open"] \the [src]!</span>")
-			if(open)
-				close(user)
-			else
-				open()
+
+/obj/structure/pit/attackby(obj/item/I, mob/user, params)
+	if(user.a_intent == INTENT_HARM)
+		return ..()
+
+	if(istype(I, /obj/item/shovel))
+		add_fingerprint(user)
+		var/prev_state = open
+		user.visible_message(
+			span_notice("[user] starts [prev_state ? "filling" : "digging open"] [src]."),
+			span_notice("You start [prev_state ? "filling" : "digging open"] [src]..."),
+		)
+		I.play_tool_sound(src)
+		if(!do_after(user, 5 SECONDS * I.toolspeed, src, category = DA_CAT_TOOL) || open != prev_state)
+			return ATTACK_CHAIN_PROCEED
+		I.play_tool_sound(src)
+		if(open)
+			close(user)
 		else
-			to_chat(user, "<span class='notice'>You stop shoveling.</span>")
-		return
-	if (!open && istype(W,/obj/item/stack/sheet/wood))
-		if(locate(/obj/structure/gravemarker) in src.loc)
-			to_chat(user, "<span class='notice'>There's already a grave marker here.</span>")
-		else
-			visible_message("<span class='notice'>\The [user] starts making a grave marker on top of \the [src]</span>")
-			if(do_after(user, 5 SECONDS * W.toolspeed, src, category = DA_CAT_TOOL))
-				visible_message("<span class='notice'>\The [user] finishes the grave marker</span>")
-				var/obj/item/stack/sheet/wood/plank = W
-				plank.use(2)
-				new/obj/structure/gravemarker(src.loc)
-			else
-				to_chat(user, "<span class='notice'>You stop making a grave marker.</span>")
-		return
-	..()
+			open()
+		user.visible_message(
+			span_notice("[user] finishes [prev_state ? "filling" : "digging open"] [src]."),
+			span_notice("You have finished [prev_state ? "filling" : "digging open"] [src]."),
+		)
+		return ATTACK_CHAIN_PROCEED_SUCCESS
+
+	if(istype(I, /obj/item/stack/sheet/wood))
+		if(open)
+			to_chat(user, span_warning("You need to fill [src] first."))
+			return ATTACK_CHAIN_PROCEED
+		if(locate(/obj/structure/gravemarker) in loc)
+			to_chat(user, span_warning("There's already the grave marker installed."))
+			return ATTACK_CHAIN_PROCEED
+		var/obj/item/stack/sheet/wood/wood = I
+		if(wood.get_amount() < 2)
+			to_chat(user, span_warning("You need at least two planks of wood to do this."))
+			return ATTACK_CHAIN_PROCEED
+		user.visible_message(
+			span_notice("[user] starts making the grave marker on top of [src]."),
+			span_notice("You start making a grave marker on top of [src]..."),
+		)
+		if(!do_after(user, 5 SECONDS * I.toolspeed, src, category = DA_CAT_TOOL) || open || (locate(/obj/structure/gravemarker) in loc) || !wood.use(2))
+			return ATTACK_CHAIN_PROCEED
+		var/obj/structure/gravemarker/gravemarker = new(loc)
+		gravemarker.add_fingerprint(user)
+		user.visible_message(
+			span_notice("[user] finishes construction of the grave marker on top of [src]."),
+			span_notice("You have constructed a grave marker on top of [src]..."),
+		)
+		return ATTACK_CHAIN_PROCEED_SUCCESS
+
+	return ..()
 
 
 /obj/structure/pit/update_icon_state()
@@ -212,15 +238,33 @@
 
 	message = "Здесь упокоен [nam], [born] - [died]."
 
-/obj/structure/gravemarker/attackby(obj/item/W, mob/user)
-	if(istype(W,/obj/item/hatchet))
-		visible_message("<span class = 'warning'>\The [user] starts hacking away at \the [src] with \the [W].</span>")
-		if(do_after(user, 3 SECONDS))
-			visible_message("<span class = 'warning'>\The [user] hacks \the [src] apart.</span>")
-			new /obj/item/stack/sheet/wood(src)
-			new /obj/item/stack/sheet/wood(src)
-			qdel(src)
-	if(is_pen(W))
-		var/msg = sanitize(input(user, "What should it say?", "Grave marker", message) as text|null)
+
+/obj/structure/gravemarker/attackby(obj/item/I, mob/user, params)
+	if(user.a_intent == INTENT_HARM)
+		return ..()
+
+	if(is_pen(I))
+		var/msg = tgui_input_text(user, "What should it say?", "Grave marker", "Rest In Peace")
 		if(msg)
 			message = msg
+		return ATTACK_CHAIN_PROCEED_SUCCESS
+
+	if(istype(I, /obj/item/hatchet))
+		user.visible_message(
+			span_notice("[user] starts hacking away [src] with [I]."),
+			span_notice("You start hacking away [src] with [I]..."),
+		)
+		I.play_tool_sound(src)
+		if(!do_after(user, 3 SECONDS * I.toolspeed, src, category = DA_CAT_TOOL))
+			return ATTACK_CHAIN_PROCEED
+		I.play_tool_sound(src)
+		user.visible_message(
+			span_notice("[user] hacks [src] apart."),
+			span_notice("You have hacked [src] apart."),
+		)
+		var/obj/item/stack/sheet/wood/wood = new(drop_location(), 2)
+		wood.add_fingerprint(user)
+		qdel(src)
+		return ATTACK_CHAIN_BLOCKED_ALL
+
+	return ..()
